[ClavaAst] Changes way to find the source node of a CFG
This fixes a bug when a loop is the only direct child statement of the
body. Since the loop had in edges, it would not be considered a source.

The current method filters the blocks BB, finding the  one whose first
stmt is the same as the body first stmt.