Add SingleFunctionReturn pass
Add pass to Clava's normalization utilities API that transforms a
function that has multiple returns into a function that has a single
return.

Does so by inserting a label at the end of the function with a single
return statement, and replaces existing returns with an assignment to a
return variable and a jump to the label.