fix(ClavaLaraApi): Use proper constructor for compound assignment creation
Compound assignments have a specialized node type that subclasses binary op. It is so because clang
computes the diferent intermediate types that might imply conversions between the operands. e.g.:

```c
int i = 1;
i += 2.3;
# type of lhs -> int
# type of rhs -> double
# intermediate type of lhs for computation -> double (implicit cast)
# intermediate type of result prior to assignment -> double
# type of expression -> int (implicit cast)
```

Therefore there is a specific factory for compound assignments, `ClavaJoinPoints.compoundAssign`,
that is now used instead of `ClavaJoinPoints.binaryOp`.
When it is a simple assignment, it also uses ClavaJoinPoints.assignment instead of `.binaryOp`.